refactor(middlewares): remove Middleware suffix from file names
