+++ server/src/middlewares/checkSearchParams.ts
