+++ server/src/middlewares/index.ts
@@ -1,4 +1,4 @@
-export * from "./authorizationTwitchMiddleware";
-export * from "./checkSearchParamsMiddleware";
-export * from "./errorHandlersMiddleware";
-export * from "./isParamObjectIdMiddleware";
+export * from "./authorizationTwitch";
+export * from "./checkSearchParams";
+export * from "./errorHandlers";
+export * from "./isParamObjectId";
